Importer: legal name change updates
