+++ search-solr-importer/src/search_solr_importer/config.py
@@ -63,12 +63,20 @@ class Config():  # pylint: disable=too-few-public-methods
 
     # POSTGRESQL
     SQLALCHEMY_TRACK_MODIFICATIONS = False
+    DB_LOCATION = os.getenv('DATABASE_LOCATION', 'OCP')
 
     DB_USER = os.getenv('DATABASE_USERNAME', '')
     DB_PASSWORD = os.getenv('DATABASE_PASSWORD', '')
     DB_NAME = os.getenv('DATABASE_NAME', '')
     DB_HOST = os.getenv('DATABASE_HOST', '')
     DB_PORT = os.getenv('DATABASE_PORT', '5432')
+
+    if DB_LOCATION == 'GCP':
+        DB_USER = os.getenv('DATABASE_USERNAME_GCP', '')
+        DB_PASSWORD = os.getenv('DATABASE_PASSWORD_GCP', '')
+        DB_NAME = os.getenv('DATABASE_NAME_GCP', '')
+        DB_HOST = os.getenv('DATABASE_HOST_GCP', '')
+        DB_PORT = os.getenv('DATABASE_PORT_GCP', '5432')
 
     if DB_UNIX_SOCKET := os.getenv('DATABASE_UNIX_SOCKET', None):
         SQLALCHEMY_DATABASE_URI = f'postgresql+psycopg2://{DB_USER}:{DB_PASSWORD}@/{DB_NAME}?host={DB_UNIX_SOCKET}'
